wip: fixed statistics issues

+++ extensions/datastores/cassandra/src/main/java/mil/nga/giat/geowave/datastore/cassandra/operations/CassandraStatisticsIterator.java
@@ -40,8 +40,8 @@ public class CassandraStatisticsIterator implements
 			}
 			else {
 				if (statEntry.getStatisticsId().equals(
-						currentStatistics.getStatisticsId()) && Short.valueOf(
-						statEntry.getInternalDataAdapterId()).equals(
+						currentStatistics.getStatisticsId()) && 
+						statEntry.getInternalDataAdapterId().equals(
 						currentStatistics.getInternalDataAdapterId())) {
 					currentStatistics.merge(statEntry);
 				}
